+++ reduction/reduction-CUB/CUB/cub/reduce/codes/cub_reduce_half.cu
@@ -109,10 +109,10 @@ __global__ void Half2FloatKernel(
 }
 
 
-float Initialize_normal(float *h_in, int num_items, int seed)
+double Initialize_normal(float *h_in, int num_items, int seed)
 {
     //srand(412);
-    float inclusive = 0.0;
+    double inclusive = 0.0;
     //std::random_device rd;
     std::mt19937 gen{seed};
     std::normal_distribution<> d{0,1};
@@ -134,15 +134,15 @@ float Initialize_normal(float *h_in, int num_items, int seed)
     return inclusive;
 }
 
-float Initialize_uniform(float *h_in, int num_items, int seed)
+double Initialize_uniform(float *h_in, int num_items, int seed)
 {  
     //srand(12);
     std::mt19937 gen(seed);
     std::uniform_real_distribution<> d(0, 1);
-    float inclusive = 0.0;
+    double inclusive = 0.0;
     for (int i = 0; i < num_items; ++i)
     {   
-        h_in[i] = (float) d(gen);
+        h_in[i] = (double) d(gen);
         //h_in[i] = (float)rand()/((float)(RAND_MAX)*1000);
         inclusive += h_in[i]; 
     }
@@ -229,7 +229,7 @@ int main(int argc, char** argv)
 
 
     // lectura de archivo en host i
-    float h_reference;
+    double h_reference;
     if (dist==0){
         h_reference = Initialize_normal(h_in, num_items,seed);
     }
